requirements.txt
fastapi
uvicorn
requests
pyppeteer
beautifulsoup4
pytest-asyncio
pypdf
pyppeteer_stealth
chromium
boto3
botocore
pytest
python-dotenv
PyMuPDF